Alias for converting MySQL select output to tab-separated fields
so I can paste the data into a spreadsheet.

To use, run a query in a MySQL console. Then pipe that output into
tabify. A crude but effective way to do this on Mac OS is to highlight
the MySQL session in the console, copy it to the clipboard, and then run

  pbpaste | tabify

Or to send the tab-separated version straight back to the clipboard,

  pbpaste | tabify | pbcopy

This will scrape only the tabular data from the MySQL session, and
convert it to tab-separated values for quick pasting into your favorite
spreadsheet application.